mcp-server/requirements.txt
mcp<1.20.0
sentence-transformers>=2.2.0
chromadb>=0.4.0
pyyaml>=6.0
